phy: msm: usb: Drive pulse on D+ data line in case of CDP
It is required to driver pulse in case of CDP detection, when
connected to CDP port of hosts. Else some hosts transition
the CDP port to SDP port by disconnecting VBUS and re-enabling
VBUS, if no pullup applied within 2sec after charger detection.
Due to this, device ends up charging with 500mA instead of
1500mA current and takes more time to charge battery fully.
Hence apply pulse on D+ data line in case of CDP to avoid this
behavior.

Change-Id: I357bc4d6598ca1deb431a896a7f6f655501bb7f8
Signed-off-by: Vijayavardhan Vennapusa <vvreddy@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>