usb: dwc3-msm: Set host mode HSPHY speed flags based on CCS
Currently the driver checks whether the PED bit of the PORTSC
register is set to decide whether any device is connected, and
then checks the PORTSPEED field to check for the speed and
accordingly updates the HSPHY speed flag with LS or FS/HS. Hence,
the PDC is programmed to fire HS_PHY_IRQ only on the rising edge
of Dp/Dm.
In a case where the connected peripheral has failed to enumerate,
the PED bit can be cleared by the host driver. When such a device
is disconnected, there is no HS_PHY_IRQ fired to wakeup SW and
disconnect handling does not happen.
Fix this by setting HSPHY speed flags based on the CCS bit of the
PORTSC register so that the PDC is programmed to fire HS_PHY_IRQ
on the falling edge of Dp/Dm.

Change-Id: I122d170434da19fd6aeb022a5fa8a85f1ea6f541
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>